fix: merapikan isi & tatak letak kode fitur login
