+++ src/components/hooks/useLogin.jsx
@@ -1,0 +1,42 @@
+import Cookies from "js-cookie";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+
+export const useLogin = () => {
+  const router = useRouter();
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+
+  async function handleSubmitLogin(event) {
+    event.preventDefault();
+
+    try {
+      const res = await fetch("https://eventmakers-api.fly.dev/auth/login", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          email,
+          password,
+        }),
+      });
+      const { payload, token } = await res.json();
+
+      // set payload to localstorage
+      localStorage.setItem("user", JSON.stringify(payload));
+
+      // set token to cookie
+      Cookies.set("token", token);
+
+      // jika berhasil login, redirect ke halaman events
+      router.push("/events");
+
+    } catch {
+      // jika gagal login, kembalikan ke halaman login
+      router.push("/");
+    }
+  }
+
+  return { handleSubmitLogin, setEmail, setPassword };
+};
